Fixing code review comments
