+++ bftengine/include/bftengine/DbCheckpointManager.hpp
@@ -13,21 +13,43 @@
 
 #pragma once
 
-#include "Serializable.h"
-#include "DbCheckpointHandler.hpp"
 #include <vector>
 #include <chrono>
 #include <PrimitiveTypes.hpp>
 #include "callback_registry.hpp"
-
+#include "status.hpp"
+#include "storage/db_interface.h"
+#include "Serializable.h"
+#include "bftengine/PersistentStorage.hpp"
+#include <optional>
+#include <string>
+#include <mutex>
+#include <queue>
+#include <condition_variable>
+#include <bftengine/DbCheckpointMetadata.hpp>
+#include "Metrics.hpp"
+#include <algorithm>
+#include <thread>
+#if __has_include(<filesystem>)
+#include <filesystem>
+namespace fs = std::filesystem;
+#elif __has_include(<experimental/filesystem>)
+#include <experimental/filesystem>
+namespace _fs = std::experimental::filesystem;
+#else
+#error "Missing filesystem support"
+#endif
 namespace bftEngine::impl {
+using std::chrono::duration_cast;
 using SeqNum = bftEngine::impl::SeqNum;
+using Status = concordUtils::Status;
+using SystemClock = std::chrono::system_clock;
+using Seconds = std::chrono::seconds;
 class DbCheckpointManager {
  public:
   void createDbCheckpoint(const SeqNum& seqNum);
-  void setLastCheckpointCreationTime(const std::chrono::seconds& lastTime) { lastCheckpointCreationTime_ = lastTime; }
   std::chrono::seconds getLastCheckpointCreationTime() const { return lastCheckpointCreationTime_; }
-  void initializeDbCheckpointHanlder(const std::shared_ptr<concord::storage::IDBClient>& dbClient,
+  void initializeDbCheckpointManager(std::shared_ptr<concord::storage::IDBClient> dbClient,
                                      std::shared_ptr<bftEngine::impl::PersistentStorage> p,
                                      std::shared_ptr<concordMetrics::Aggregator> aggregator,
                                      const std::function<uint64_t()>& getLastBlockIdCb);
@@ -37,16 +59,55 @@ class DbCheckpointManager {
     static DbCheckpointManager instance_;
     return instance_;
   }
+  ~DbCheckpointManager() {
+    stopped_ = true;
+    cv_.notify_one();
+    if (cleanupThread_.joinable()) cleanupThread_.join();
+  }
 
  private:
   logging::Logger getLogger() {
     static logging::Logger logger_(logging::getLogger("concord.bft.db_checkpoint_manager"));
     return logger_;
   }
-  DbCheckpointManager() = default;
-  std::chrono::seconds lastCheckpointCreationTime_{0};
-  std::unique_ptr<concord::storage::IDbCheckPointHandler> dbCheckpointHandler_;
+  DbCheckpointManager()
+      : metrics_{concordMetrics::Component("rocksdbCheckpoint", std::make_shared<concordMetrics::Aggregator>())},
+        maxDbCheckpointCreationTimeMsec_(metrics_.RegisterGauge("maxDbCheckpointCreationTimeInMsecSoFar", 0)),
+        dbCheckpointCreationAverageTimeMsec_(metrics_.RegisterGauge("dbCheckpointCreationAverageTimeMsec", 0)),
+        lastDbCheckpointSizeInMb_(metrics_.RegisterGauge("lastDbCheckpointSizeInMb", 0)),
+        lastDbCheckpointBlockId_(metrics_.RegisterGauge("lastDbCheckpointBlockId", 0)),
+        numOfDbCheckpointsCreated_(metrics_.RegisterCounter("numOfDbCheckpointsCreated", 0)) {
+    metrics_.Register();
+  }
+  void init();
+  Status createDbCheckpoint(const uint64_t& checkPointId, const uint64_t& lastBlockId, const uint64_t& seqNum);
+  void removeCheckpoint(const uint64_t& checkPointId);
+  void removeAllCheckpoints() const;
+  void cleanUp();
   std::function<uint64_t()> getLastBlockIdCb_;
+  // get total size recursively
+  uint64_t directorySize(const _fs::path& directory, const bool& excludeHardLinks);
+  // get checkpoint metadata
+  void loadCheckpointDataFromPersistence();
+  void checkforCleanup();
+  bool stopped_{false};
+  DbCheckpointMetadata dbCheckptMetadata_;
+  std::shared_ptr<concord::storage::IDBClient> dbClient_;
+  std::shared_ptr<bftEngine::impl::PersistentStorage> ps_;
+  std::queue<CheckpointId> checkpointToBeRemoved_;
+  std::thread cleanupThread_;
+  std::condition_variable cv_;
+  std::mutex lock_;
+  uint32_t maxNumOfCheckpoints_{0};  // 0-disabled
+  uint64_t lastCheckpointSeqNum_{0};
+  std::chrono::seconds lastCheckpointCreationTime_{duration_cast<Seconds>(SystemClock::now().time_since_epoch())};
+  std::string dbCheckPointDirPath_;
+  concordMetrics::Component metrics_;
+  concordMetrics::GaugeHandle maxDbCheckpointCreationTimeMsec_;
+  concordMetrics::GaugeHandle dbCheckpointCreationAverageTimeMsec_;
+  concordMetrics::GaugeHandle lastDbCheckpointSizeInMb_;
+  concordMetrics::GaugeHandle lastDbCheckpointBlockId_;
+  concordMetrics::CounterHandle numOfDbCheckpointsCreated_;
 };
 
 }  // namespace bftEngine::impl
